requirements.txt
django
django-rest-framework
djangorestframework-simplejwt
drf-yasg
python-dotenv
psycopg2-binary